🚚 Move the TabOverview file to the new location
