+++ modules/ppcp-settings/resources/js/Components/Screens/Settings/Tabs/index.js
@@ -1,6 +1,6 @@
 import { __ } from '@wordpress/i18n';
 
-import TabOverview from '../../Overview/TabOverview';
+import TabOverview from './TabOverview';
 import TabPaymentMethods from '../../Overview/TabPaymentMethods';
 import TabSettings from '../../Overview/TabSettings';
 import TabStyling from '../../Overview/TabStyling';
